Package data provider is storing negative size values

+++ src/data_provider/src/packages/appxWindowsWrapper.h
@@ -95,7 +95,7 @@ class AppxWindowsWrapper final : public IPackageWrapper
             return UNKNOWN_VALUE;
         }
 
-        int size() const override
+        int64_t size() const override
         {
             return 0;
         }
